+++ packages/easter-day/test/calculator.d.ts
@@ -1,0 +1,2 @@
+declare function calculator(y: number): Date
+export = calculator